Fix Play2 friendly tasks #213
