+++ task/src/main/scala/skinny/task/generator/ReverseModelAllGenerator.scala
@@ -12,17 +12,21 @@ object ReverseModelAllGenerator extends ReverseModelAllGenerator {
 
 trait ReverseModelAllGenerator extends CodeGenerator {
 
-  private[this] def showUsage = {
+  protected def showUsage = {
     showSkinnyGenerator()
-    println("""  Usage: sbt "task/run generate:reverse-model-all [skinnyEnv]""")
+    println("""  Usage: sbt "task/run generate:reverse-model-all [env]""")
     println("")
+  }
+
+  protected def initializeDB(skinnyEnv: Option[String]): Unit = {
+    System.setProperty(SkinnyEnv.PropertyKey, skinnyEnv.getOrElse(SkinnyEnv.Development))
+    DBSettings.initialize()
   }
 
   def run(args: List[String]) {
     val skinnyEnv: Option[String] = args.headOption
 
-    System.setProperty(SkinnyEnv.PropertyKey, skinnyEnv.getOrElse(SkinnyEnv.Development))
-    DBSettings.initialize()
+    initializeDB(skinnyEnv)
 
     val tables: Seq[Table] = DB.getAllTableNames.filter(_.toLowerCase != "schema_version").flatMap { tableName =>
       DB.getTable(tableName)
